+++ Extensions.cs
@@ -16,8 +16,10 @@ namespace DictionaryHelper
 		/// </summary>
 		/// <param name="key">Dictionary key</param>
 		/// <param name="defaultValue">Optional: If dictionary item is not found and has empty value then default value will be returned</param>
+		/// <param name="parentKey">Optional: Key will be created as a child of the parent key.</param>
 		/// <param name="create">Optional: If key does not exist it will be created. Requires default value.</param>
-		public static string DictionaryValue(this UmbracoHelper helper, string key, string defaultValue = null, bool create = false)
+		/// <returns>Dictionary string value</returns>
+		public static string DictionaryValue(this UmbracoHelper helper, string key, string defaultValue = null, string parentKey = null, bool create = false)
 		{
 			var content = helper.UmbracoContext.PublishedContentRequest;
 			string culture = Thread.CurrentThread.CurrentCulture.Name;
@@ -27,7 +29,7 @@ namespace DictionaryHelper
 				culture = content.Culture.Name;
 			}
 
-			var value = Service.GetValueByKeyAndCulture(key, culture, defaultValue, create);
+			var value = Service.GetValueByKeyAndCulture(key, culture, defaultValue, parentKey, create);
 
 			return value;
 		}