+++ Models/DictionaryItem.cs
@@ -11,5 +11,6 @@ namespace DictionaryHelper.Models
 		public string Key { get; set; }
 		public string Value { get; set; }
 		public string Culture { get; set; }
+		public Guid Id { get; set; }
 	}
 }