+++ Service.cs
@@ -7,11 +7,13 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using Umbraco.Core;
+using Umbraco.Core.Services;
 
 namespace DictionaryHelper
 {
 	public static class Service
 	{
+		private static IEnumerable<Umbraco.Core.Models.ILanguage> _allLanguages = null;
 
 		public static IEnumerable<DictionaryItem> GetAll()
 		{
@@ -23,8 +25,26 @@ namespace DictionaryHelper
 			return DictionaryCache._cache.Any(x => x.Value.Key == key);
 		}
 
-		public static DictionaryItem GetByKeyAndCulture(string key, string culture, string defaultValue = null, bool create = false)
+		public static DictionaryItem GetDictionaryItem(string key)
 		{
+			if (KeyExist(key))
+			{
+				return DictionaryCache._cache.FirstOrDefault(x => x.Value.Key == key).Value;
+			}
+
+			return null;
+		}
+
+		public static DictionaryItem GetByKeyAndCulture(string key, string culture, string defaultValue = null, string parentKey = null, bool create = false)
+		{
+			var keys = new string[] {};
+
+			if (key.Contains("."))
+			{
+				keys = key.Split('.');
+
+				key = keys.Last();
+			}
 
 			if (DictionaryCache._cache.Any(x => x.Value.Key == key && x.Value.Culture == culture)) {
 
@@ -37,43 +57,46 @@ namespace DictionaryHelper
 
 			}
 
-			if (!string.IsNullOrEmpty(defaultValue))
+			if (create)
 			{
-				if (create)
+				var ls = ApplicationContext.Current.Services.LocalizationService;
+
+				_allLanguages = ls.GetAllLanguages();
+
+				if (keys.Length > 0)
 				{
-					var ls = ApplicationContext.Current.Services.LocalizationService;
+					var item = CreateDictionaryTree(keys, defaultValue, culture);
 
-					var languages = ls.GetAllLanguages();
+					return item;
+				} else
+				{
+					DictionaryItem parentItem = null;
 
-					var language = languages.FirstOrDefault(x => x.IsoCode == culture);
-
-					if (language != null)
+					if (!string.IsNullOrEmpty(parentKey))
 					{
-						var dict = ls.CreateDictionaryItemWithIdentity(key, null, defaultValue);
-
-						ls.AddOrUpdateDictionaryValue(dict, language, defaultValue);
-
-						DictionaryCache.AddOrUpdate(key, culture, defaultValue);
-
-						ls.Save(dict);
+						parentItem = GetDictionaryItem(parentKey);
 					}
 
-				}
+					var item = CreateDictionaryItem(key, defaultValue, parentItem != null ? parentItem.Id : (Guid?)null, culture);
 
+					return item;
+				}
+			} else
+			{
 				return new DictionaryItem()
 				{
+					Id = Guid.Empty,
 					Culture = culture,
 					Key = key,
 					Value = defaultValue
 				};
 			}
 
-			return null;
 		}
 
-		public static string GetValueByKeyAndCulture(string key, string culture, string defaultValue = null, bool create = false)
+		public static string GetValueByKeyAndCulture(string key, string culture, string defaultValue = null, string parentKey = null, bool create = false)
 		{
-			var dict = GetByKeyAndCulture(key, culture, defaultValue, create);
+			var dict = GetByKeyAndCulture(key, culture, defaultValue, parentKey, create);
 
 			if (dict != null)
 			{
@@ -82,5 +105,89 @@ namespace DictionaryHelper
 
 			return string.Empty;
 		}
+
+		private static DictionaryItem CreateDictionaryTree(string[] keys, string defaultValue, string culture)
+		{
+
+			for (int i = 0; i < keys.Length; i++)
+			{
+				var key = keys[i];
+
+				var item = GetDictionaryItem(key);
+
+				if (item == null)
+				{
+
+					DictionaryItem parentItem = null;
+					var _defaultValue = string.Empty;
+
+					if (i > 0)
+					{
+						parentItem = GetDictionaryItem(keys[i-1]);
+					}
+
+					if (i == keys.Length - 1)
+					{
+						_defaultValue = defaultValue;
+					}
+
+					var dict = CreateDictionaryItem(key, _defaultValue, parentItem != null ? parentItem.Id : (Guid?)null, culture);
+
+					if (i == keys.Length - 1)
+					{
+						return dict;
+					}
+
+				} else
+				{
+					if (i == keys.Length - 1)
+					{
+						return item;
+					}
+				}
+			}
+
+			return null;
+
+		}
+
+		private static DictionaryItem CreateDictionaryItem(string key, string defaultValue, Guid? parent, string culture)
+		{
+			var language = _allLanguages.FirstOrDefault(x => x.IsoCode == culture);
+
+			if (language != null)
+			{
+				var ls = ApplicationContext.Current.Services.LocalizationService;
+
+				var dict = ls.CreateDictionaryItemWithIdentity(key, parent, defaultValue);
+
+				foreach (var la in _allLanguages)
+				{
+					UpdateDictionaryItemCache(ls, dict, la, defaultValue);
+				}
+
+				ls.Save(dict);
+
+				return new DictionaryItem()
+				{
+					Culture = culture,
+					Id = dict.Key,
+					Key = key,
+					Value = defaultValue
+				};
+			}
+
+			return null;
+		}
+
+		private static void UpdateDictionaryItemCache(ILocalizationService ls, Umbraco.Core.Models.IDictionaryItem dict, Umbraco.Core.Models.ILanguage language, string defaultValue)
+		{
+			ls.AddOrUpdateDictionaryValue(dict, language, defaultValue);
+		}
+
+		private static readonly ILog Log =
+				LogManager.GetLogger(
+					MethodBase.GetCurrentMethod().DeclaringType
+				);
 	}
 }