+++ Changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.3] - 2017-12-22
+
+- Feature: Dictionary can be created under specific parent key
+- Feature: Dictionary tree can be created by using for example "root.child.key". This will create 3 dictionary keys.
+- Fix: Delete event, will remove all children from cache correctly
+- Fix: Save event runs only once instead of twice.
+- fix: Culture name not set correctly when adding to cache after save event
+
 ## [1.0.2] - 2017-12-22
 
 - Fixed framework nuget issues.